requirements.txt
pip
autopep8
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-Migrate==4.0.4
PyMySQL==1.1.0
python-dotenv==1.0.0
Flask-JWT-Extended==4.5.2
bcrypt==4.0.1
marshmallow==3.20.1
gunicorn==22.0.0
Werkzeug==3.0.6